[RF] Remove RooAbsRootFinder
