+++ roofit/roofitcore/inc/LinkDef.h
@@ -64,7 +64,6 @@
 #pragma link C++ class RooAbsProxy+ ;
 #pragma link C++ class RooAbsReal+ ;
 #pragma link C++ class RooAbsRealLValue+ ;
-#pragma link C++ class RooAbsRootFinder+ ;
 #pragma link C++ class RooAddGenContext+ ;
 #pragma link C++ class RooAddition+ ;
 #pragma link C++ class RooAddModel+ ;
